refactor: 모달 구조 변경 및 설정
